% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effects.R
\name{add_autoregressive_cross_lagged}
\alias{add_autoregressive_cross_lagged}
\title{add_autoregressive_cross_lagged}
\usage{
add_autoregressive_cross_lagged(internal_list)
}
\arguments{
\item{internal_list}{internal list with info_variables}
}
\value{
data.frame with autoregressive and cross-lagged effects
}
\description{
Creates autoregressive and cross-lagged effects for the observed variables
}
\keyword{internal}
