% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effects.R
\name{add_homogeneous_covariances}
\alias{add_homogeneous_covariances}
\title{add_homogeneous_covariances}
\usage{
add_homogeneous_covariances(internal_list)
}
\arguments{
\item{internal_list}{internal list with info_variables}
}
\value{
data.frame with covariances
}
\description{
Creates variances between the lagged time varying variables in a
homogeneous model and covariances between contemporaneous variables
}
\keyword{internal}
