% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effects.R
\name{add_latent_residual}
\alias{add_latent_residual}
\title{add_latent_residual}
\usage{
add_latent_residual(internal_list)
}
\arguments{
\item{internal_list}{internal list with info_variables}
}
\value{
data.frame with latent traits
}
\description{
Creates latent traits for each of the observed variables
}
\keyword{internal}
