% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effects.R
\name{add_observed_exogenous_covariances}
\alias{add_observed_exogenous_covariances}
\title{add_observed_exogenous_covariances}
\usage{
add_observed_exogenous_covariances(internal_list)
}
\arguments{
\item{internal_list}{internal list}
}
\value{
data.frame with covariances
}
\description{
Add covariances between observed exogenous variables if use_definition_variables = FALSE. These
can be computed from the observed data.
}
\keyword{internal}
