% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effects.R
\name{add_process_residual_variances}
\alias{add_process_residual_variances}
\title{add_process_residual_variances}
\usage{
add_process_residual_variances(internal_list)
}
\arguments{
\item{internal_list}{internal list with info_variables}
}
\value{
data.frame with residual variances
}
\description{
Creates residual variances for the observed variables
}
\keyword{internal}
