% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effects.R
\name{add_product_covariances}
\alias{add_product_covariances}
\title{add_product_covariances}
\usage{
add_product_covariances(internal_list)
}
\arguments{
\item{internal_list}{internal list}
}
\value{
data.frame with covariances
}
\description{
Add covariances between exogenous product terms and x1, x2, ..., y1, y2, ... if use_definition_variables = FALSE
}
\keyword{internal}
