% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_product_terms.R
\name{add_product_terms}
\alias{add_product_terms}
\title{add_product_terms}
\usage{
add_product_terms(internal_list)
}
\arguments{
\item{internal_list}{internal list}
}
\value{
internal_list
}
\description{
add product terms of time varying variables and time invariant variables
in case of nonlinear models. The product terms will be added to
internal_list$info_data$data with the naming convention x3z1.
}
\keyword{internal}
