% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_starting_values.R
\name{add_starting_values}
\alias{add_starting_values}
\title{Add Starting Values to Model}
\usage{
add_starting_values(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted list with several slots in \code{..$info_data} filled
in.
}
\description{
Add starting values to an existing model.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492,
 DOI: 10.1080/10705511.2020.1780598
}
\keyword{external}
