% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effects.R
\name{add_time_invariant_predictors}
\alias{add_time_invariant_predictors}
\title{add_time_invariant_predictors}
\usage{
add_time_invariant_predictors(internal_list)
}
\arguments{
\item{internal_list}{internal list with info_variables}
}
\value{
data.frame with exogenous predictors
}
\description{
Creates exogenous predictors as specified in internal_list$info_variables$info_time_invariant_variables
}
\keyword{internal}
