% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_panel_sem.R
\name{check_panel_sem_specification}
\alias{check_panel_sem_specification}
\title{check_panel_sem_specification}
\usage{
check_panel_sem_specification(specification)
}
\arguments{
\item{specification}{list with user specified arguments}
}
\value{
throws error in case of misspecification
}
\description{
checks if the user specified all arguments of fit_panel_sem correctly
}
\keyword{internal}
