% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_list.R
\name{create_empty_list}
\alias{create_empty_list}
\title{Create Empty Internal List}
\usage{
create_empty_list(verbose)
}
\arguments{
\item{verbose}{Integer number describing verbosity of console output.
0...no output (default); 1...user messages; 2...debugging-relevant messages.}
}
\value{
An empty internal list with the following structure:\cr
 \tabular{lll}{
    List of XX\cr
   \code{$info_data}: List of XX       \tab \tab  \cr
     \code{..$data}: \code{data.frame}    \tab \tab User-specified data set.\cr
     \code{..$data_product_terms}: \code{data.frame}    \tab \tab Data set with product terms of observed variables. \cr
     \code{..$n_obs}: \code{int(0)}   \tab \tab Number of observations.\cr
     \code{..$n_var}: \code{int(0)}   \tab \tab Total number of variables in the data set.\cr
     \code{..$var_names}: \code{chr(0)}  \tab \tab Names of variables.\cr
   \code{$info_model}: List of XX        \tab \tab \cr
     \code{..$n_ocassions}: \code{int(0)}  \tab \tab Number of measurement occasions. \cr
     \code{..$n_processes}: \code{int(0)}  \tab \tab Number of dynamic processes. \cr
     \code{..$n_time_invariant} \code{int(0)}  \tab \tab  Number of observed time-invariant variables \cr
     \code{..$linear} \code{logical(0)}  \tab \tab \code{TRUE} if the model is linear in observed variables.\cr
     \code{..$heterogeneity} \code{chr(0)}  \tab \tab Specify type of unobserved heterogeneity.\cr
     \code{..$use_resamples} \code{logical(0)}  \tab \tab . If \code{TRUE}, a resampling procedure for starting values and model diagnostics is performed.\cr
     \code{..$use_open_mx} \code{logical(0)}  \tab \tab If \code{TRUE}, \code{OpenMx} is used, otherwise \code{lavaan} is used.\cr
   \code{$info_variables}: List of XX        \tab \tab \cr
     \code{..$user_names_time_varying}: \code{char[0,0]}   \tab \tab User-Specified names of observed time-varying variables. \cr
     \code{..$generic_names_time_varying}: \code{char[0,0]}   \tab \tab Generic names of observed time-varying variables. \cr
     \code{..$names_processes}: \code{char[0,0]}   \tab \tab Names of dynamic processes. \cr
     \code{..$info_time_invariant_varianbles}: \code{char[0,0]}   \tab \tab Names and causal structure of observed time-invariant variables. \cr
     \code{..$names_time_invariant_unique}: \code{char[0,0]}   \tab \tab Names of unique observed time-invariant variables. \cr
     \code{..$names_time_invariant_unobserved_additive}: \code{char[0,0]}   \tab \tab Names of UNobserved time-invariant variables. \cr
     \code{..$names_time_invariant_unobserved_autoregressive}: \code{char[0,0]}   \tab \tab Names of UNobserved time-invariant variables. \cr
     \code{..$names_time_invariant_unobserved_cross_lagged}: \code{char[0,0]}   \tab \tab Names of UNobserved time-invariant variables. \cr
   \code{$info_parameters}: List of XX \tab \tab \cr
     \code{..$C_table}: \code{data.frame} \tab \tab Table with labeled model parameters (structural coefficients).\cr
     \code{..$Psi_table}: \code{data.frame} \tab \tab Table with labeled model parameters (covariance matrix).\cr
   \code{model_matrices}: List of XX \tab \tab \cr
    \code{..$C_labels}: \code{char[0,0]}  \tab \tab Character matrix of labels of structural coefficients. \cr
    \code{..$Psi_labels}: \code{char[0,0]}  \tab \tab Character matrix of labels of covariance matrix. \cr
    \code{..$select_observed_only}: \code{num[0,0]}  \tab \tab Numeric matrix to select observed variables. \cr
    \code{..$C_start}: \code{num[0,0]}  \tab \tab Numeric matrix of starting values of structural coefficients. \cr
    \code{..$Psi_start}: \code{num[0,0]}  \tab \tab Numeric matrix of starting values of covariance matrix. \cr
  \code{$model_syntax}: List of XX \tab \tab \cr
     \code{..$lavaan}: \code{list} \tab \tab Model specified in \code{lavaan}. \cr
     \code{..$OpenMx}: \code{list} \tab \tab Model specified in \code{OpenMx}. \cr
   \code{$control}: List of XX     \tab \tab \cr
     \code{..$verbose}: \code{int(0)}     \tab \tab Verbosity of console output.\cr
   \code{$tables}: List of XX      \tab \tab \cr
     \code{..$summary}: \code{data.frame}  \tab \tab Information for \code{\link{summary.panelSEM}} function.\cr
     \code{..$print}: \code{data.frame}  \tab \tab Information for \code{\link{print.panelSEM}}  function.\cr
     \code{..$coef}: \code{data.frame}  \tab \tab Information for \code{\link{coef.panelSEM}}  function.
 }
}
\description{
Creates an empty internal list of predefined structure. The
list will subsequently be filled while the function
\code{\link{fit_panel_sem}} is running.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492,
 DOI: 10.1080/10705511.2020.1780598
}
