% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_values.R
\name{create_latent_proxy}
\alias{create_latent_proxy}
\title{create_latent_proxy}
\usage{
create_latent_proxy(residualized_data, user_names_time_varying, latent_traits)
}
\arguments{
\item{residualized_data}{residualized data}

\item{user_names_time_varying}{matrix with names of time varying variables}

\item{latent_traits}{vector with names of the latent traits}
}
\value{
matrix with proxy for latent traits
}
\description{
Computes the rowmeans of the data for all processes using the
time varying variables (except for the first occasion). These rowmeans
are then used as proxy for the latent traits
}
\keyword{internal}
