% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_long_data.R
\name{create_long_data}
\alias{create_long_data}
\title{create_long_data}
\usage{
create_long_data(data, time_varying_variables, time_invariant_variables)
}
\arguments{
\item{data}{data set}

\item{time_varying_variables}{List of character vectors containing names of the time-varying
variables. The number entries in the list corresponds to the number of univariate time-series.
Each character vector contains the time-ordered variable names of a univariate time-series
starting with the first measurement occasion.}

\item{time_invariant_variables}{List of character vectors containing names of the time-invariant
variables. List must have the same length as list in argument \code{time_varying_variables}.
@keywords internal}
}
\description{
changes data set to be in long format, including lagged effects for
all time_varying_variables.
}
