% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_matrix.R
\name{create_model_matrix}
\alias{create_model_matrix}
\title{Create Model Matrices}
\usage{
create_model_matrix(internal_list = NULL, matrix_type = NULL)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}

\item{matrix_type}{Character string describing the type of model matrix. Admissible values are: \code{"C"}, \code{"Psi"}, and \code{"selection"}.}
}
\value{
A matrix of model parameters or a selection matrix.
}
\description{
Create model matrices such as the matrix of structural
coefficients, the covariance matrix, and selection matrix.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492,
 DOI: 10.1080/10705511.2020.1780598
}
