% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_panelSEM_s3_object.R
\name{create_panelSEM_s3_object}
\alias{create_panelSEM_s3_object}
\title{panelSEM S3 Object}
\usage{
create_panelSEM_s3_object(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
Object of class \code{panelSEM}.
}
\description{
Assign class \code{panelSEM} to internal list.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z
}
