% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_table.R
\name{create_parameter_table}
\alias{create_parameter_table}
\title{Create Parameter Table}
\usage{
create_parameter_table(internal_list, parameter_type)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}

\item{parameter_type}{Character string describing the type of model matrix. Admissible values are: \code{"C"}, \code{"Psi"}.}
}
\value{
A table of model parameters.
}
\description{
Create a table of model parameters.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492,
 DOI: 10.1080/10705511.2020.1780598
}
