% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_RAM_matrices.R
\name{create_RAM_matrices}
\alias{create_RAM_matrices}
\title{create_RAM_matrices}
\usage{
create_RAM_matrices(parameter_table)
}
\arguments{
\item{parameter_table}{data frame with parameters}
}
\value{
list with C (directed effects), Psi (undirected effects),
M (intercepts), and F (filter) matrices
}
\description{
create empty RAM matrices
}
\keyword{internal}
