% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_model.R
\name{.check_input}
\alias{.check_input}
\title{.check_input}
\usage{
.check_input(
  data,
  labels_time_varying_variables,
  labels_time_invariant_variables
)
}
\arguments{
\item{data}{data set in wide format}

\item{labels_time_varying_variables}{vector with labels of time - varying variables}

\item{labels_time_invariant_variables}{list with vectors holding labels of time - invariant variables}
}
\value{
throws error
}
\description{
check input to specify_model function
}
