% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_model.R
\name{.get_time_varying_variable_base}
\alias{.get_time_varying_variable_base}
\title{.get_time_varying_variable_base}
\usage{
.get_time_varying_variable_base(labels_time_varying_variables)
}
\arguments{
\item{labels_time_varying_variables}{vector with labels of time - varying variables}
}
\value{
vector with labels
}
\description{
returns just the names of time-varying variables without the time indices
}
