% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_estimates.R
\name{fill_in_estimates}
\alias{fill_in_estimates}
\title{Fit Panel Data Model}
\usage{
fill_in_estimates(internal_list, rand = FALSE)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}

\item{rand}{Logical value indicating if (\code{TRUE}) random starting values
should be chosen.}
}
\value{
An object of class \code{panelSEM} for which several methods
are available including \code{\link{summary.panelSEM}} and
\code{\link{print.panelSEM}}.
}
\description{
Fit a model from the class of dynamic panel data models to longitudinal data.
Models include linear and nonlinear cross-lagged panel models with additive or nonadditive
unobserved heterogeneity.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492,
 DOI: 10.1080/10705511.2020.1780598
}
\keyword{external}
