% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_info_variables.R
\name{fill_in_info_variables}
\alias{fill_in_info_variables}
\title{Get Information about Variables}
\usage{
fill_in_info_variables(
  internal_list,
  time_varying_variables,
  time_invariant_variables,
  linear,
  heterogeneity,
  use_open_mx,
  use_definition_variables
)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}

\item{time_varying_variables}{List of character vectors containing names of the time-varying
variables. Within each vector, the variable names must be time-ordered starting with the first
measurement ocassion.}

\item{time_invariant_variables}{List of character vectors containing names of the time-invariant
variables. List must have the same length as list in argument \code{time_varying_variables}.}

\item{linear}{Logical (TRUE = default / FALSE) indicating if the model is linear in observed variables (TRUE).}

\item{heterogeneity}{Character vector indicating the type of unobserved heterogeneity. Admissible values are \code{"homogeneous"}, \code{"additive"}, \code{"autoregressive"}, and \code{"cross-lagged"} (or any non-conflicting combination).}

\item{use_open_mx}{Logical (TRUE / FALSE default) indicating if \code{lavaan} (FALSE) or \code{OpenMx} (TRUE)
should be used.}

\item{use_definition_variables}{When setting linear = FALSE, there are two ways to add interactions
between the previous observation and the time invariant predictors (e.g., y2 = (c_y_x + c_y_x_z2*z2)*x1 + ...)).
First, definition variables can be used (currently only supported by OpenMx). Second, product terms
can be added explicitly. When setting use_definition_variables = FALSE, panelSEM will compute the product terms (e.g., z2*x1).
These will then be used to predict the time varying variables (e.g., y2). Additionally, panelSEM will add
all covariances between product terms (e.g., prod_z2_x1 ~~ prod_z3_x1), product terms and exogenous predictors (e.g., prod_z2_x1 ~~ z3),
and product terms and time varying variables (e.g., prod_z2_x1 ~~ y5). This will result in fairly large models and long run times.}
}
\value{
The inputted internal_list with several slots filled in:
\tabular{lll}{
 \code{..$n_ocassions}: \code{int(0)}  \tab \tab Number of measurement occasions. \cr
 \code{..$n_processes}: \code{int(0)}  \tab \tab Number of dynamic processes. \cr
 \code{..$generic_names_time_varying}: \code{data.frame} \tab \tab Table with generic variable names of time-varying variables.\cr
 \code{..$user_names_time_varying}: \code{data.frame} \tab \tab Table with user-specified variable names of time-varying variables.\cr
 \code{..$generic_names_time_invariant}: \code{data.frame} \tab \tab Table with generic variable names of time-invariant variables.\cr
 \code{..$user_names_time_invariant}: \code{data.frame} \tab \tab Table with user-specified variable names of time-invariant variables.}
}
\description{
Provide information about the variables in the dynamic panel data model.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492,
 DOI: 10.1080/10705511.2020.1780598
}
