% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_starting_values.R
\name{fill_in_starting_values}
\alias{fill_in_starting_values}
\title{Compute Starting Values}
\usage{
fill_in_starting_values(internal_list)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}
}
\value{
The inputted list with the column starting values filled in to the
parameter following lists: \code{..$info_parameters$parameter_list_C} and
\code{..$info_parameters$parameter_list_Psi}.
}
\description{
Compute starting values based on simplified auxiliary models
(e.g., models without latent variables) which are similar to the target model
of interest.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492,
 DOI: 10.1080/10705511.2020.1780598
}
