% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_RAM_matrices.R
\name{fill_RAM_matrices}
\alias{fill_RAM_matrices}
\title{fill_RAM_matrices}
\usage{
fill_RAM_matrices(parameter_table)
}
\arguments{
\item{parameter_table}{data frame with parameters}
}
\value{
list with C (directed effects), Psi (undirected effects),
M (intercepts), and F (filter) matrices
}
\description{
create RAM matrices based on parameter table
}
\keyword{internal}
