% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_common_substring.R
\name{find_common_substring}
\alias{find_common_substring}
\title{find_common_substring}
\usage{
find_common_substring(strings)
}
\arguments{
\item{strings}{vector with strings}
}
\value{
string
}
\description{
returns the largest common substring of a vector of strings.
}
\keyword{internal}
