% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_values.R
\name{find_latent_traits}
\alias{find_latent_traits}
\title{find_latent_traits}
\usage{
find_latent_traits(parameter_table, user_names_time_varying)
}
\arguments{
\item{parameter_table}{data.frame with parameters}

\item{user_names_time_varying}{matrix with names of time varying variables}
}
\value{
vector with names of latent traits for each of the processes
}
\description{
Extracts the names of the latent traits for each of the processes. The initial
occasion is not taken into account.
}
\keyword{internal}
