% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\name{fit_nonlinear_nonadditive}
\alias{fit_nonlinear_nonadditive}
\title{fit_nonlinear_nonadditive}
\usage{
fit_nonlinear_nonadditive(
  data,
  time_varying_variables,
  time_invariant_variables,
  use_resamples = FALSE,
  verbose = 0
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data with named columns. Data MUST be in wide format. If set to NULL,
only the lavaan or OpenMx model will be returned. This model can be used to simulate data or check the
specification without providing data.}

\item{time_varying_variables}{List of character vectors containing names of the time-varying
variables. The number entries in the list corresponds to the number of univariate time-series.
Each character vector contains the time-ordered variable names of a univariate time-series
starting with the first measurement occasion.}

\item{time_invariant_variables}{List of character vectors containing names of the time-invariant
variables. List must have the same length as list in argument \code{time_varying_variables}.}

\item{use_resamples}{Logical (TRUE / FALSE = default) indicating if a resampling procedure is used for the computation of starting values and model diagnostics.}

\item{verbose}{Integer number describing the verbosity of console output.
Admissible values: 0: no output (default), 1: user messages,
2: debugging-relevant messages.}
}
\value{
An object of class \code{panelSEM} for which several methods
are available including \code{\link{summary.panelSEM}} and
\code{\link{print.panelSEM}}.
}
\description{
Fit a non-linear model with additive and cross-lagged heterogeneity to longitudinal data.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492,
 DOI: 10.1080/10705511.2020.1780598
}
