% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_panel_sem.R
\name{fit_panel_sem}
\alias{fit_panel_sem}
\title{Fit Dynamic Panel Data Models}
\usage{
fit_panel_sem(
  data = NULL,
  time_varying_variables = NULL,
  time_invariant_variables = NULL,
  homogeneous = FALSE,
  linear = TRUE,
  additive = TRUE,
  use_resamples = FALSE,
  use_open_mx = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data with named columns. Data MUST be in wide format.}

\item{time_varying_variables}{List of character vectors containing names of the time-varying
variables. The number entries in the list corresponds to the number of univariate time-series.
Each character vector contains the time-ordered variable names of a univariate time-series 
starting with the first measurement occasion.}

\item{time_invariant_variables}{List of character vectors containing names of the time-invariant
variables. List must have the same length as list in argument \code{time_varying_variables}.}

\item{homogeneous}{Logical (TRUE / FALSE = default) indicating if the model contains unobserved heterogeneity (TRUE).}

\item{linear}{Logical (TRUE = default / FALSE) indicating if the model is linear in observed variables (TRUE).}

\item{additive}{Logical (TRUE = default / FALSE) indicating if the unobserved heterogeneity is additive (TRUE).}

\item{use_resamples}{Logical (TRUE / FALSE = default) indicating if a resampling procedure is used for the computation of starting values and model diagnostics.}

\item{use_open_mx}{Logical (TRUE / FALSE default) indicating if \code{lavaan} (FALSE) or \code{OpenMx} (TRUE) 
should be used.}

\item{verbose}{Integer number describing the verbosity of console output.
Admissible values: 0: no output (default), 1: user messages, 
2: debugging-relevant messages.}
}
\value{
An object of class \code{panelSEM} for which several methods 
are available including \code{\link{summary.panelSEM}} and 
\code{\link{print.panelSEM}}.
}
\description{
Fit a model from the class of dynamic panel data models to longitudinal data. 
Models include linear and nonlinear cross-lagged panel models with additive or nonadditive
unobserved heterogeneity.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible 
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87, 
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural 
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492, 
 DOI: 10.1080/10705511.2020.1780598
}
\keyword{external}
