% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_RAM_matrices.R
\name{get_variables}
\alias{get_variables}
\title{get_variables}
\usage{
get_variables(parameter_table)
}
\arguments{
\item{parameter_table}{data frame with parameters}
}
\value{
list with latents and manifests
}
\description{
Get the manifest and latent variables
}
\keyword{internal}
