% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_logger.R
\name{initialize_logger}
\alias{initialize_logger}
\title{initialize_logger}
\usage{
initialize_logger()
}
\value{
nothing
}
\description{
Create a logger layout for the logger R package
}
\keyword{internal}
