% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tested_settings.R
\name{list_append}
\alias{list_append}
\title{list_append}
\usage{
list_append(my_list, element)
}
\arguments{
\item{my_list}{list to which the element should be added}

\item{element}{element to add}
}
\value{
list with added element
}
\description{
Adds an element at the end of a list
}
\keyword{internal}
