% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{print.panelSEM}
\alias{print.panelSEM}
\title{print a panelSEM object}
\usage{
print.panelSEM(x, ...)
}
\arguments{
\item{x}{object of class panelSEM}

\item{...}{not used}
}
\value{
nothing
}
\description{
print a panelSEM object
}
