% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_values.R
\name{residualize_time_varying}
\alias{residualize_time_varying}
\title{residualize_time_varying}
\usage{
residualize_time_varying(
  user_names_time_varying,
  info_time_invariant_variables,
  data
)
}
\arguments{
\item{user_names_time_varying}{matrix with names of user defined time varying
variables}

\item{info_time_invariant_variables}{list with names of user defined time invariant
variables}

\item{data}{data.frame with data}
}
\value{
data.frame with residualized data
}
\description{
Predicts all time varying variables with the time invariant predictors and
creates a new data set containing only the residuals of these predictions.
}
\keyword{internal}
