% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_starting_values.R
\name{set_starting_values}
\alias{set_starting_values}
\title{set_starting_values}
\usage{
set_starting_values(internal_list)
}
\arguments{
\item{internal_list}{internal model list object}
}
\value{
OpenMx model or lavaan parameter table
}
\description{
changes the starting values of the model object. In case of OpenMx, an OpenMx
model is returned, where the starting values are set to those indicated in
internal_list$info_parameters$parameter_table$start. In case of lavaan, a
parameter table is returned that specifies the starting values in the ustart
column
}
\keyword{internal}
