% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_starting_values.R
\name{set_starting_values_lavaan}
\alias{set_starting_values_lavaan}
\title{set_starting_values_lavaan}
\usage{
set_starting_values_lavaan(model_syntax, starting_values)
}
\arguments{
\item{model_syntax}{lavaan model syntax}

\item{starting_values}{labeled vector with values of the starting values}
}
\value{
a lavaan parameter table with starting values
}
\description{
change the starting values of a lavaan model object.
}
\examples{
# Example copied from ?lavaan::sem
library(lavaan)
model <- '
  # latent variable definitions
     ind60 =~ x1 + x2 + x3
     dem60 =~ y1 + a*y2 + b*y3 + c*y4
     dem65 =~ y5 + a*y6 + b*y7 + c*y8

  # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60

  # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
'

fit <- sem(model, data = PoliticalDemocracy)

# extract current estimates
starting_values <- coef(fit)
# change values
starting_values[] <- rnorm(length(starting_values))^2

new_pt <- set_starting_values_lavaan(model_syntax    = model,
                                     starting_values = starting_values)
# print new values
new_pt
}
