% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_starting_values.R
\name{set_starting_values_openmx}
\alias{set_starting_values_openmx}
\title{set_starting_values_openmx}
\usage{
set_starting_values_openmx(model, starting_values)
}
\arguments{
\item{model}{mxModel object}

\item{starting_values}{labeled vector with values of the starting values}
}
\description{
change the starting values of an OpenMx model object
}
\examples{
# Example copied from ?OpenMx::mxRun
library(OpenMx)
data(demoOneFactor)

manifests <- names(demoOneFactor)
latents   <- c("G")

model <- mxModel(type="RAM",
                 manifestVars = names(demoOneFactor),
                 latentVars   = "G",
                 mxPath(from=latents, to=manifests, labels=paste("b", 1:5, sep="")),
                 mxPath(from=manifests, arrows=2, labels=paste("u", 1:5, sep="")),
                 mxPath(from=latents  , arrows=2, free=FALSE, values=1.0),
                 mxData(cov(demoOneFactor), type="cov", numObs=500))
model <- mxRun(model)

# get the current parameter values
starting_values <- omxGetParameters(model)
# change the values
starting_values[] <- rnorm(length(starting_values))^2

# set starting values
model_start <- set_starting_values_openmx(model           = model,
                                          starting_values = starting_values)
print(omxGetParameters(model_start))
print(starting_values)
}
