% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_values.R
\name{starting_values_initial_occasion}
\alias{starting_values_initial_occasion}
\title{starting_values_initial_occasion}
\usage{
starting_values_initial_occasion(
  data,
  user_names_time_varying,
  info_time_invariant_variables,
  latent_proxy,
  parameter_table
)
}
\arguments{
\item{data}{data.frame with data}

\item{user_names_time_varying}{matrix with names of user defined time varying
variables}

\item{info_time_invariant_variables}{list with names of user defined time invariant
variables}

\item{latent_proxy}{matrix with proxy variables for latent traits}

\item{parameter_table}{parameter table}
}
\value{
parameter_table with added starting values for initial occasion
}
\description{
Create starting values for the initial occasion
}
\keyword{internal}
