% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_values.R
\name{starting_values_trait_variances}
\alias{starting_values_trait_variances}
\title{starting_values_trait_variances}
\usage{
starting_values_trait_variances(latent_proxy, parameter_table)
}
\arguments{
\item{latent_proxy}{matrix with proxy variables for latent traits}

\item{parameter_table}{parameter table}
}
\value{
parameter_table with added starting values for latent trait variances
and covariances
}
\description{
Create starting values for the latent traits.
}
\keyword{internal}
