% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{summary.panelSEM}
\alias{summary.panelSEM}
\title{Summarize a panelSEM object}
\usage{
summary.panelSEM(object, ...)
}
\arguments{
\item{object}{object of class panelSEM}

\item{...}{not used}
}
\value{
nothing
}
\description{
Summarize a panelSEM object
}
