% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_data.R
\name{fill_in_data}
\alias{fill_in_data}
\title{fill_in_data}
\usage{
fill_in_data(data = NULL, internal_list = NULL, add_product_variables = FALSE)
}
\arguments{
\item{data}{A data.frame containing the data in the wide format.}

\item{internal_list}{A list with various information extracted from the
model.}

\item{add_product_variables}{Logical value indicating if (\code{TRUE})
product terms of observed variables should be added to the
\code{..$info_data} slot.}
}
\value{
The inputted list with several slots in \code{..$info_data} filled
in.
}
\description{
Fill Data Into List
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492,
 DOI: 10.1080/10705511.2020.1780598
}
