% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_info_variables.R
\name{fill_in_info_variables}
\alias{fill_in_info_variables}
\title{Get Information about Variables}
\usage{
fill_in_info_variables(
  internal_list = NULL,
  time_varying_variables = NULL,
  time_invariant_variables = NULL,
  linear = TRUE,
  heterogeneity = NULL,
  use_open_mx = FALSE
)
}
\arguments{
\item{internal_list}{A list with various information extracted from the
model.}

\item{time_varying_variables}{List of character vectors containing names of the time-varying
variables. Within each vector, the variable names must be time-ordered starting with the first
measurement ocassion.}

\item{time_invariant_variables}{List of character vectors containing names of the time-invariant
variables. List must have the same length as list in argument \code{time_varying_variables}.}

\item{linear}{Logical (TRUE = default / FALSE) indicating if the model is linear in observed variables (TRUE).}

\item{heterogeneity}{Character vector indicating the type of unobserved heterogeneity. Admissible values are \code{"homogeneous"}, \code{"additive"}, \code{"autoregressive"}, and \code{"cross-lagged"} (or any non-conflicting combination).}

\item{use_open_mx}{Logical (TRUE / FALSE default) indicating if \code{lavaan} (FALSE) or \code{OpenMx} (TRUE)
should be used.}

\item{n_occasions}{Integer number indicating the number of measurement occasions.}
}
\value{
The inputted internal_list with several slots filled in:
\tabular{lll}{
 \code{..$n_ocassions}: \code{int(0)}  \tab \tab Number of measurement occasions. \cr
 \code{..$n_processes}: \code{int(0)}  \tab \tab Number of dynamic processes. \cr
 \code{..$generic_names_time_varying}: \code{data.frame} \tab \tab Table with generic variable names of time-varying variables.\cr
 \code{..$user_names_time_varying}: \code{data.frame} \tab \tab Table with user-specified variable names of time-varying variables.\cr
 \code{..$generic_names_time_invariant}: \code{data.frame} \tab \tab Table with generic variable names of time-invariant variables.\cr
 \code{..$user_names_time_invariant}: \code{data.frame} \tab \tab Table with user-specified variable names of time-invariant variables.}
}
\description{
Provide information about the variables in the dynamic panel data model.
}
\references{
Gische, C., Voelkle, M.C. (2022) Beyond the Mean: A Flexible
Framework for Studying Causal Effects Using Linear Models. Psychometrika 87,
868–901. https://doi.org/10.1007/s11336-021-09811-z

Gische, C., West, S.G., & Voelkle, M.C. (2021) Forecasting Causal
 Effects of Interventions versus Predicting Future Outcomes, Structural
 Equation Modeling: A Multidisciplinary Journal, 28:3, 475-492,
 DOI: 10.1080/10705511.2020.1780598
}
