% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_model_specification.R
\name{fill_in_model_specification}
\alias{fill_in_model_specification}
\title{fill_in_model_specification}
\usage{
fill_in_model_specification(internal_list)
}
\arguments{
\item{internal_list}{internal list object}
}
\value{
internal_list, where lavaan or OpenMx model is added to internal_list$model_syntax
}
\description{
Creates a lavaan or OpenMx model based in the user specification in internal_list
}
